An initial ChromeDriver test to load a page.
The unfortunate piece of this is it looks like there is no way to override
in the .NET Selenium Webdriver the name of the executable for the Chrome
Driver.

The constructor for ChromeDriver takes a path to the directory that the
"chromedriver" executable lives in. Giving this the path to where my
executable lived did not work. This is because my executable was loaded
by the selenium-standalone npm package, and the chromedriver executable is
named "2.25-x64-chromedriver".

https://www.npmjs.com/package/selenium-standalone

I was not under the influence of wanting to change this. For me it lives at
/usr/local/lib/node_modules/selenium-standalone/.selenium/chromedriver/

and as it was in a git repo, I didn't feel it wise to rename.

So, what does one do?

I copied it to the directory

/Users/corywheeler/Documents/projects/chromestuff/

So that the full path is now

/Users/corywheeler/Documents/projects/chromestuff/chromedriver

where "chromedriver" is a copy of "2.25-x64-chromedriver".

I have yet to find a way to be able to set the name of the chrome driver
executable name, which is unfortunate. Perhaps there is one, I've just
not stumbled onto it yet.